grid and blank rename
