+++ templates/Web/Pages/Angular.Grid/src/app/app-shell/AngularGrid/grid-box/grid-box.component.css
