+++ templates/Web/Pages/Angular.Grid/src/app/app-shell/AngularGrid/grid.component.ts
@@ -1,6 +1,6 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { CONSTANTS } from '../../../constants';
-import { IGridTextItem } from './GridBox/grid-box.component';
+import { IGridTextItem } from './grid-box/grid-box.component';
 import { GridService } from './grid.service';
 import { DomSanitizer } from '@angular/platform-browser';
 import { MatIconRegistry } from '@angular/material/icon';
